fix(relations): issue with belongsToMany save method
Before saving a relationship in a pivot table it was attempting to save the first model. This was

returning false if the model wasn't dirty and didn't need saving, then silently failing to save the

pivot. I have changed this to a try catch, so if the save returns 0 it still goes on to insert the

pivot record

Closes #107